Refactor: Simplify Nebula Effect and Reorganize Code
- I extracted the complex `EnhancedNebula` component and its helper classes (`AdaptiveQualityManager`, `MobileGPUComputationRenderer`) from `App.jsx` into their own separate files in `src/components` and `src/lib`.

- I simplified the main application by replacing the `EnhancedNebula` component with the simpler and more self-contained `InteractiveNebula` component in `App.jsx`.

- This change improves code organization, reduces the complexity of the main App component, and fulfills your request for a functional, modern React implementation of the nebula effect.
