+++ dev_server.log
@@ -1,0 +1,9 @@
+
+> blaize-it@1.0.0 dev
+> vite
+
+
+  VITE v5.4.19  ready in 320 ms
+
+  ➜  Local:   http://localhost:5173/
+  ➜  Network: use --host to expose
