+++ src/App.jsx
@@ -1,582 +1,6 @@
-import React, { useState, useEffect, useRef, Suspense, useMemo } from "react";
+import React, { useState, useEffect, useRef, Suspense } from "react";
 import { Menu, X, CheckCircle, XCircle, Loader2, ChevronLeft, ChevronRight } from 'lucide-react';
-import * as THREE from 'three';
-
-// Enhanced Adaptive Quality Manager with advanced mobile optimizations
-class AdaptiveQualityManager {
-  constructor(renderer, scene) {
-    this.renderer = renderer;
-    this.scene = scene;
-    this.qualitySettings = {
-      pixelRatio: Math.min(window.devicePixelRatio, 2), // Cap at 2x for performance
-      particleCount: this.getInitialParticleCount(),
-      rayMarchSteps: this.getInitialRayMarchSteps(),
-      postProcessing: true,
-    };
-    this.performanceMonitor = {
-      frameTimes: [],
-      stable: true,
-      gpuUtilization: 0,
-      lastFrameTime: performance.now(),
-    };
-    this.isLowEnd = this.detectLowEndDevice();
-    this.memoryBudget = this.calculateMemoryBudget();
-  }
-
-  detectLowEndDevice() {
-    const canvas = document.createElement('canvas');
-    const gl = canvas.getContext('webgl');
-    if (!gl) return true;
-
-    const debugInfo = gl.getExtension('WEBGL_debug_renderer_info');
-    const renderer = debugInfo ? gl.getParameter(debugInfo.UNMASKED_RENDERER_WEBGL) : '';
-
-    // Detect low-end GPUs
-    const lowEndGPUs = ['powervr', 'adreno 3', 'adreno 4', 'mali-4', 'intel'];
-    return lowEndGPUs.some(gpu => renderer.toLowerCase().includes(gpu)) ||
-           navigator.hardwareConcurrency < 4;
-  }
-
-  getInitialParticleCount() {
-    if (this.isLowEnd) return 25000;
-    if (window.innerWidth < 768) return 50000;
-    return 100000;
-  }
-
-  getInitialRayMarchSteps() {
-    return this.isLowEnd ? 16 : 32;
-  }
-
-  calculateMemoryBudget() {
-    // Conservative estimate: 2MB per million screen pixels
-    const screenPixels = window.innerWidth * window.innerHeight;
-    return Math.floor(screenPixels * 2 / 1000000); // MB
-  }
-
-  update() {
-    const now = performance.now();
-    const frameTime = now - this.performanceMonitor.lastFrameTime;
-    this.performanceMonitor.lastFrameTime = now;
-
-    this.performanceMonitor.frameTimes.push(frameTime);
-    if (this.performanceMonitor.frameTimes.length > 60) {
-      this.performanceMonitor.frameTimes.shift();
-    }
-
-    const avgFrameTime = this.performanceMonitor.frameTimes.reduce((a, b) => a + b, 0) / this.performanceMonitor.frameTimes.length;
-
-    // GPU bound detection (simplified)
-    if (avgFrameTime > 33) { // Slower than 30fps
-      this.reduceRenderQuality();
-      this.performanceMonitor.stable = false;
-    } else if (avgFrameTime < 13 && this.performanceMonitor.frameTimes.length >= 30) { // Faster than 75fps
-      this.increaseQuality();
-      this.performanceMonitor.stable = true;
-    }
-  }
-
-  reduceRenderQuality() {
-    // Reduce pixel ratio first
-    if (this.qualitySettings.pixelRatio > 0.5) {
-      this.qualitySettings.pixelRatio = Math.max(this.qualitySettings.pixelRatio * 0.9, 0.5);
-      this.renderer.setPixelRatio(this.qualitySettings.pixelRatio);
-      return;
-    }
-
-    // Then reduce particle count
-    if (this.qualitySettings.particleCount > 10000) {
-      this.qualitySettings.particleCount *= 0.8;
-    }
-
-    // Disable post-processing
-    this.qualitySettings.postProcessing = false;
-  }
-
-  increaseQuality() {
-    if (this.qualitySettings.pixelRatio < Math.min(window.devicePixelRatio, 2)) {
-      this.qualitySettings.pixelRatio = Math.min(this.qualitySettings.pixelRatio * 1.05, Math.min(window.devicePixelRatio, 2));
-      this.renderer.setPixelRatio(this.qualitySettings.pixelRatio);
-    }
-  }
-}
-
-// Enhanced GPU Computation Renderer with mobile optimizations
-class MobileGPUComputationRenderer {
-  constructor(sizeX, sizeY, renderer) {
-    this.variables = [];
-    this.currentTextureIndex = 0;
-    this.sizeX = sizeX;
-    this.sizeY = sizeY;
-    this.renderer = renderer;
-
-    this.scene = new THREE.Scene();
-    this.camera = new THREE.Camera();
-    this.camera.position.z = 1;
-
-    // Use HalfFloat on mobile for better performance
-    this.dataType = this.isMobile() ? THREE.HalfFloatType : THREE.FloatType;
-
-    this.passThruUniforms = {
-      passThruTexture: { value: null }
-    };
-
-    this.init();
-  }
-
-  isMobile() {
-    return /(iPad|iPhone|iPod|Android)/g.test(navigator.userAgent);
-  }
-
-  addResolutionDefine(material) {
-    material.defines.resolution = `vec2(${this.sizeX.toFixed(1)}, ${this.sizeY.toFixed(1)})`;
-  }
-
-  createShaderMaterial(computeFragmentShader, uniforms = {}) {
-    const material = new THREE.ShaderMaterial({
-      uniforms: uniforms,
-      vertexShader: this.getPassThroughVertexShader(),
-      fragmentShader: computeFragmentShader
-    });
-
-    this.addResolutionDefine(material);
-    return material;
-  }
-
-  init() {
-    const passThruShader = this.createShaderMaterial(this.getPassThroughFragmentShader(), this.passThruUniforms);
-    const mesh = new THREE.Mesh(new THREE.PlaneGeometry(2, 2), passThruShader);
-    this.scene.add(mesh);
-    this.mesh = mesh;
-    this.passThruShader = passThruShader;
-  }
-
-  addVariable(variableName, computeFragmentShader, initialValueTexture) {
-    const material = this.createShaderMaterial(computeFragmentShader);
-
-    const variable = {
-      name: variableName,
-      initialValueTexture: initialValueTexture,
-      material: material,
-      renderTargets: [],
-      wrapS: THREE.ClampToEdgeWrapping,
-      wrapT: THREE.ClampToEdgeWrapping,
-      minFilter: THREE.NearestFilter,
-      magFilter: THREE.NearestFilter
-    };
-
-    this.variables.push(variable);
-
-    variable.renderTargets[0] = this.createRenderTarget();
-    variable.renderTargets[1] = this.createRenderTarget();
-    this.renderTexture(initialValueTexture, variable.renderTargets[0]);
-    this.renderTexture(initialValueTexture, variable.renderTargets[1]);
-
-    return variable;
-  }
-
-  createRenderTarget() {
-    return new THREE.WebGLRenderTarget(this.sizeX, this.sizeY, {
-      wrapS: THREE.ClampToEdgeWrapping,
-      wrapT: THREE.ClampToEdgeWrapping,
-      minFilter: THREE.NearestFilter,
-      magFilter: THREE.NearestFilter,
-      format: THREE.RGBAFormat,
-      type: this.dataType,
-      stencilBuffer: false
-    });
-  }
-
-  createTexture() {
-    const size = this.sizeX * this.sizeY * 4;
-    const data = this.dataType === THREE.HalfFloatType ?
-      new Uint16Array(size) : new Float32Array(size);
-
-    return new THREE.DataTexture(data, this.sizeX, this.sizeY, THREE.RGBAFormat, this.dataType);
-  }
-
-  compute() {
-    const currentTextureIndex = this.currentTextureIndex;
-    const nextTextureIndex = this.currentTextureIndex === 0 ? 1 : 0;
-
-    for (let v = 0; v < this.variables.length; v++) {
-      const variable = this.variables[v];
-      const uniforms = variable.material.uniforms;
-
-      // Set dependencies
-      for (let d = 0; d < this.variables.length; d++) {
-        const depVar = this.variables[d];
-        if (uniforms.hasOwnProperty(depVar.name)) {
-          uniforms[depVar.name].value = depVar.renderTargets[currentTextureIndex].texture;
-        }
-      }
-
-      this.doRenderTarget(variable.material, variable.renderTargets[nextTextureIndex]);
-    }
-
-    this.currentTextureIndex = nextTextureIndex;
-  }
-
-  getCurrentRenderTarget(variable) {
-    return variable.renderTargets[this.currentTextureIndex];
-  }
-
-  doRenderTarget(material, output) {
-    this.mesh.material = material;
-    this.renderer.setRenderTarget(output);
-    this.renderer.render(this.scene, this.camera);
-    this.mesh.material = this.passThruShader;
-  }
-
-  renderTexture(input, output) {
-    this.passThruUniforms.passThruTexture.value = input;
-    this.doRenderTarget(this.passThruShader, output);
-    this.passThruUniforms.passThruTexture.value = null;
-  }
-
-  getPassThroughVertexShader() {
-    return `
-      void main() {
-        gl_Position = vec4(position, 1.0);
-      }
-    `;
-  }
-
-  getPassThroughFragmentShader() {
-    return `
-      uniform sampler2D passThruTexture;
-      void main() {
-        vec2 uv = gl_FragCoord.xy / resolution.xy;
-        gl_FragColor = texture2D(passThruTexture, uv);
-      }
-    `;
-  }
-}
-
-// Enhanced Nebula Component with advanced shader techniques
-function EnhancedNebula({ adaptiveQualityManager }) {
-  const containerRef = useRef();
-  const sceneRef = useRef();
-  const rendererRef = useRef();
-  const mouseRef = useRef([0, 0]);
-  const timeRef = useRef(0);
-
-  // Mobile-optimized shaders with FBM and domain warping
-  const positionFragmentShader = `
-    uniform float u_time;
-    uniform sampler2D u_velocities;
-    uniform float u_stepScale;
-
-    void main() {
-      vec2 uv = gl_FragCoord.xy / resolution.xy;
-      vec3 pos = texture2D(u_positions, uv).xyz;
-      vec3 vel = texture2D(u_velocities, uv).xyz;
-
-      // Apply velocity with adaptive step scaling for mobile
-      pos += vel * 0.016 * u_stepScale;
-
-      // Boundary conditions
-      if (length(pos.xy) > 8.0) {
-        pos.xy = normalize(pos.xy) * 8.0;
-      }
-
-      gl_FragColor = vec4(pos, 1.0);
-    }
-  `;
-
-  const velocityFragmentShader = `
-    uniform float u_time;
-    uniform vec2 u_mouse;
-    uniform float u_noiseScale;
-
-    // High-performance FBM with rotation matrix (from research)
-    const mat2 m2 = mat2(0.8, -0.6, 0.6, 0.8);
-
-    float noise(vec2 p) {
-      return fract(sin(dot(p, vec2(12.9898, 78.233))) * 43758.5453) * 2.0 - 1.0;
-    }
-
-    float fbm(vec2 p) {
-      float f = 0.0;
-      f += 0.5000 * noise(p); p = m2 * p * 2.02;
-      f += 0.2500 * noise(p); p = m2 * p * 2.03;
-      f += 0.1250 * noise(p); p = m2 * p * 2.01;
-      f += 0.0625 * noise(p);
-      return f / 0.9375;
-    }
-
-    // Domain warping for organic patterns (from research)
-    float pattern(vec2 p) {
-      vec2 q = vec2(fbm(p + vec2(0.0, 0.0)),
-                    fbm(p + vec2(5.2, 1.3)));
-
-      vec2 r = vec2(fbm(p + 4.0 * q + vec2(1.7, 9.2)),
-                    fbm(p + 4.0 * q + vec2(8.3, 2.8)));
-
-      return fbm(p + 4.0 * r);
-    }
-
-    void main() {
-      vec2 uv = gl_FragCoord.xy / resolution.xy;
-      vec3 pos = texture2D(u_positions, uv).xyz;
-      vec3 vel = texture2D(u_velocities, uv).xyz;
-
-      // Mouse interaction with distance-based influence
-      float dist = distance(pos.xy, u_mouse);
-      float influence = 1.0 - smoothstep(0.0, 3.0, dist);
-      vec2 direction = normalize(pos.xy - u_mouse);
-      vel.xy += direction * influence * 0.15;
-
-      // Add organic noise using domain warping
-      vec2 noiseForce = vec2(
-        pattern(pos.xy * 0.1 + u_time * 0.02),
-        pattern(pos.xy * 0.1 + u_time * 0.02 + 100.0)
-      ) * 0.03 * u_noiseScale;
-
-      vel.xy += noiseForce;
-
-      // Damping
-      vel *= 0.98;
-
-      gl_FragColor = vec4(vel, 1.0);
-    }
-  `;
-
-  const particleVertexShader = `
-    uniform sampler2D u_positions;
-    uniform float u_size;
-    uniform float u_time;
-
-    attribute vec2 a_uv;
-
-    void main() {
-      vec3 pos = texture2D(u_positions, a_uv).xyz;
-
-      vec4 mvPosition = modelViewMatrix * vec4(pos, 1.0);
-      gl_Position = projectionMatrix * mvPosition;
-
-      // Size attenuation based on distance
-      float sizeAttenuation = 300.0 / length(mvPosition.xyz);
-      gl_PointSize = u_size * sizeAttenuation;
-    }
-  `;
-
-  const particleFragmentShader = `
-    uniform float u_time;
-
-    void main() {
-      // Create circular particles with soft edges
-      vec2 center = gl_PointCoord - vec2(0.5);
-      float dist = length(center);
-
-      if (dist > 0.5) discard;
-
-      float alpha = 1.0 - smoothstep(0.0, 0.5, dist);
-      alpha *= 0.8;
-
-      // Color gradient based on position and time
-      vec3 color1 = vec3(0.2, 0.8, 1.0); // Cyan
-      vec3 color2 = vec3(0.8, 0.4, 1.0); // Purple
-      vec3 color = mix(color1, color2, sin(u_time * 0.001) * 0.5 + 0.5);
-
-      gl_FragColor = vec4(color, alpha);
-    }
-  `;
-
-  useEffect(() => {
-    if (!containerRef.current) return;
-
-    const container = containerRef.current;
-    const width = container.clientWidth;
-    const height = container.clientHeight;
-
-    // Initialize renderer with mobile optimizations
-    const renderer = new THREE.WebGLRenderer({
-      antialias: false, // Disable for mobile performance
-      alpha: true,
-      powerPreference: "high-performance"
-    });
-    renderer.setSize(width, height);
-    renderer.setPixelRatio(Math.min(window.devicePixelRatio, 2));
-    renderer.setClearColor(0x000000, 0);
-    container.appendChild(renderer.domElement);
-    rendererRef.current = renderer;
-
-    // Scene setup
-    const scene = new THREE.Scene();
-    const camera = new THREE.PerspectiveCamera(75, width / height, 0.1, 1000);
-    camera.position.z = 5;
-    sceneRef.current = scene;
-
-    // Initialize adaptive quality manager
-    if (!adaptiveQualityManager.current) {
-      adaptiveQualityManager.current = new AdaptiveQualityManager(renderer, scene);
-    }
-
-    // GPU computation setup
-    const particleCount = adaptiveQualityManager.current.qualitySettings.particleCount;
-    const textureWidth = Math.sqrt(particleCount);
-    const textureHeight = textureWidth;
-
-    const gpuCompute = new MobileGPUComputationRenderer(textureWidth, textureHeight, renderer);
-
-    // Initialize textures
-    const posTexture = gpuCompute.createTexture();
-    const velTexture = gpuCompute.createTexture();
-
-    const posArray = posTexture.image.data;
-    const velArray = velTexture.image.data;
-
-    // Initialize particle positions in a spiral galaxy pattern
-    for (let i = 0; i < posArray.length; i += 4) {
-      const angle = Math.random() * Math.PI * 2;
-      const radius = Math.sqrt(Math.random()) * 6;
-      const spiralOffset = radius * 0.2;
-
-      posArray[i] = Math.cos(angle + spiralOffset) * radius;
-      posArray[i + 1] = Math.sin(angle + spiralOffset) * radius;
-      posArray[i + 2] = (Math.random() - 0.5) * 1;
-      posArray[i + 3] = 1.0;
-
-      velArray[i] = 0;
-      velArray[i + 1] = 0;
-      velArray[i + 2] = 0;
-      velArray[i + 3] = 1.0;
-    }
-
-    posTexture.needsUpdate = true;
-    velTexture.needsUpdate = true;
-
-    // Add variables to GPU compute
-    const positionVariable = gpuCompute.addVariable("u_positions", positionFragmentShader, posTexture);
-    const velocityVariable = gpuCompute.addVariable("u_velocities", velocityFragmentShader, velTexture);
-
-    // Set up variable dependencies
-    gpuCompute.variables.forEach(variable => {
-      gpuCompute.variables.forEach(dependency => {
-        if (variable !== dependency) {
-          variable.material.uniforms[dependency.name] = { value: null };
-        }
-      });
-    });
-
-    // Add uniforms
-    positionVariable.material.uniforms.u_time = { value: 0 };
-    positionVariable.material.uniforms.u_stepScale = { value: 1.0 };
-
-    velocityVariable.material.uniforms.u_time = { value: 0 };
-    velocityVariable.material.uniforms.u_mouse = { value: new THREE.Vector2(0, 0) };
-    velocityVariable.material.uniforms.u_noiseScale = { value: 1.0 };
-
-    // Create particle geometry
-    const particleGeometry = new THREE.BufferGeometry();
-    const uvArray = new Float32Array(particleCount * 2);
-
-    for (let i = 0; i < particleCount; i++) {
-      uvArray[i * 2] = (i % textureWidth) / textureWidth;
-      uvArray[i * 2 + 1] = Math.floor(i / textureWidth) / textureHeight;
-    }
-
-    particleGeometry.setAttribute('a_uv', new THREE.BufferAttribute(uvArray, 2));
-
-    // Create particle material
-    const particleMaterial = new THREE.ShaderMaterial({
-      uniforms: {
-        u_positions: { value: null },
-        u_size: { value: adaptiveQualityManager.current.isLowEnd ? 1.5 : 2.0 },
-        u_time: { value: 0 }
-      },
-      vertexShader: particleVertexShader,
-      fragmentShader: particleFragmentShader,
-      transparent: true,
-      blending: THREE.AdditiveBlending,
-      depthWrite: false
-    });
-
-    const particles = new THREE.Points(particleGeometry, particleMaterial);
-    scene.add(particles);
-
-    // Mouse tracking
-    const handleMouseMove = (event) => {
-      const rect = container.getBoundingClientRect();
-      const x = ((event.clientX - rect.left) / rect.width) * 2 - 1;
-      const y = -((event.clientY - rect.top) / rect.height) * 2 + 1;
-      mouseRef.current = [x * 4, y * 4];
-    };
-
-    const handleResize = () => {
-      const newWidth = container.clientWidth;
-      const newHeight = container.clientHeight;
-      camera.aspect = newWidth / newHeight;
-      camera.updateProjectionMatrix();
-      renderer.setSize(newWidth, newHeight);
-    };
-
-    container.addEventListener('mousemove', handleMouseMove);
-    window.addEventListener('resize', handleResize);
-
-    // Animation loop
-    const animate = () => {
-      if (!renderer || !scene || !camera) return;
-
-      timeRef.current += 0.016;
-
-      // Update adaptive quality manager
-      if (adaptiveQualityManager.current) {
-        adaptiveQualityManager.current.update();
-
-        // Apply quality settings
-        const quality = adaptiveQualityManager.current.qualitySettings;
-        positionVariable.material.uniforms.u_stepScale.value = quality.rayMarchSteps / 32;
-        velocityVariable.material.uniforms.u_noiseScale.value = quality.postProcessing ? 1.0 : 0.5;
-      }
-
-      // Update uniforms
-      positionVariable.material.uniforms.u_time.value = timeRef.current;
-      velocityVariable.material.uniforms.u_time.value = timeRef.current;
-      velocityVariable.material.uniforms.u_mouse.value.set(mouseRef.current[0], mouseRef.current[1]);
-
-      particleMaterial.uniforms.u_time.value = timeRef.current;
-
-      // Compute GPU simulation
-      gpuCompute.compute();
-
-      // Update particle positions
-      particleMaterial.uniforms.u_positions.value = gpuCompute.getCurrentRenderTarget(positionVariable).texture;
-
-      renderer.render(scene, camera);
-      requestAnimationFrame(animate);
-    };
-
-    animate();
-
-    // Cleanup
-    return () => {
-      container.removeEventListener('mousemove', handleMouseMove);
-      window.removeEventListener('resize', handleResize);
-      if (container.contains(renderer.domElement)) {
-        container.removeChild(renderer.domElement);
-      }
-      renderer.dispose();
-      particleGeometry.dispose();
-      particleMaterial.dispose();
-    };
-  }, []);
-
-  return (
-    <div
-      ref={containerRef}
-      style={{
-        position: 'fixed',
-        top: 0,
-        left: 0,
-        width: '100%',
-        height: '100%',
-        zIndex: 0,
-        pointerEvents: 'none',
-      }}
-    />
-  );
-}
+import InteractiveNebula from './components/InteractiveNebula';
 
 // Utility Components
 function FadeInWhenVisible({ children, className = '', delay = 0, duration = 700 }) {
@@ -1215,7 +639,6 @@ function ContactSection() {
 // Main App Component
 export default function App() {
   const [currentPage, setCurrentPage] = useState("home");
-  const adaptiveQualityManager = useRef(null);
 
   useEffect(() => {
     const section = document.getElementById("home");
@@ -1268,11 +691,10 @@ export default function App() {
       `}</style>
 
       <Suspense fallback={<div className="fixed inset-0 bg-black flex items-center justify-center text-white">Loading WebGL...</div>}>
-        <EnhancedNebula adaptiveQualityManager={adaptiveQualityManager} />
+        <InteractiveNebula />
       </Suspense>
 
       <Navbar currentPage={currentPage} setCurrentPage={setCurrentPage} />
-      <PerformanceMetrics adaptiveQualityManager={adaptiveQualityManager} />
 
       <main className="pt-16">
         <HeroSection />
